# Optimize the computation of the polynomials