--- a/physics/geopotential.hpp
+++ b/physics/geopotential.hpp
@@ -47,10 +47,13 @@ class Geopotential {
   struct SurfaceFrame;
   static const Vector<double, SurfaceFrame> x_;
   static const Vector<double, SurfaceFrame> y_;
-  static const Vector<Quotient<Acceleration, GravitationalParameter>, Frame>
-      zero_;
 
   using UnitVector = Vector<double, Frame>;
+
+  template<int size>
+  using Accelerations =
+      std::array<Vector<Quotient<Acceleration, GravitationalParameter>, Frame>,
+                 size>;
 
   // Holds precomputed data for one evaluation of the acceleration.
   template<int size>