```
Run on (4 X 3310 MHz CPU s)
09/29/18 13:50:36
--------------------------------------------------------------------
Benchmark                             Time           CPU Iterations
--------------------------------------------------------------------
BM_ComputeGeopotentialCpp/2      237411 ns     236407 ns      11218
BM_ComputeGeopotentialCpp/3      308071 ns     307691 ns       8974
BM_ComputeGeopotentialCpp/5      541798 ns     541501 ns       5128
BM_ComputeGeopotentialCpp/10    3134201 ns    3147846 ns        897
BM_ComputeGeopotentialF90/2       73050 ns      73121 ns      38189
BM_ComputeGeopotentialF90/3      110732 ns     110460 ns      25280
BM_ComputeGeopotentialF90/5      200146 ns     199987 ns      13807
BM_ComputeGeopotentialF90/10     672758 ns     669003 ns       4174
```